Create Pageviews.py
A python script to fetch the pageviews of articles in a particular language wikipedia. So run that create a list of wikipedia articles, each article in a line and then save it in the same folder of the script. Here for example, "articles_eng.txt" is used. the output would be in the wikitable format.